fixing compiler errors in vs2005
git-svn-id: http://crashrpt.googlecode.com/svn@964 821a4644-a21a-0410-aec7-6199d47568cc

+++ branches/1.2/processing/crashrptprobe/MinidumpReader.cpp
@@ -638,7 +638,7 @@ int CMiniDumpReader::StackWalk(DWORD dwThreadId)
     {
       size_t pos = sSourceFile.ReverseFind('\\');
       if(pos>=0)
-        sSourceFile = sSourceFile.Mid(pos+1);
+        sSourceFile = sSourceFile.Mid((int)pos+1);
       str += _T(" [ ");
       str += sSourceFile;
       str += _T(": ");
